+++ src/components/Detail/ConfirmDeletePopup.jsx
@@ -1,9 +1,52 @@
-const ConfirmDeletePopup = (props) => {
+import axios from "axios"
+import { getJwt } from "../../utility/jwt"
+import { useNavigate, useParams } from "react-router"
+
+const ENV = import.meta.env
+const API_URL =
+  ENV.MODE === "development" ? ENV.VITE_DEV_API_URL : ENV.VITE_PROD_API_URL
+
+const ConfirmDeletePopup = ({
+  showBox,
+  onClose,
+  setErrorMessage,
+  pokemon,
+  fetchDetails,
+}) => {
+  console.log("pokemon From Pokemoncard= ", pokemon)
+  const { id } = useParams()
+  const navigate = useNavigate()
+
+  const deletePokemon = async () => {
+    try {
+      if (pokemon) {
+        const jwt = getJwt()
+        const res = await axios.delete(
+          `${API_URL}/pokemons/${pokemon.id || id}`,
+          {
+            headers: {
+              Authorization: `Bearer ${jwt}`,
+            },
+          }
+        )
+        console.log("DELETE res", res)
+      } else {
+        console.log("no data to delete")
+      }
+      // go back to list page
+      navigate("/")
+    } catch (err) {
+      setErrorMessage(JSON.stringify(err.message))
+      console.error(err)
+    } finally {
+      await fetchDetails()
+    }
+  }
   return (
-    props.showBox && (
+    showBox && (
       <div
         className="fixed inset-0 flex items-center justify-center bg-black bg-opacity-25 backdrop-blur-sm"
-        onClick={props.onClose}
+        onClick={onClose}
       >
         <div className="box-border absolute items-center p-8 m-auto bg-white rounded shadow-md w-96">
           <h2 className="text-lg">
@@ -12,13 +55,13 @@ const ConfirmDeletePopup = (props) => {
           <div className="flex justify-center mt-3">
             <button
               className="px-3 ml-2 text-lg border border-black rounded-full"
-              onClick={props.onClose}
+              onClick={onClose}
             >
               Close
             </button>
             <button
               className="px-3 ml-2 text-lg bg-red-400 border border-black rounded-full"
-              onClick={props.onDelete}
+              onClick={deletePokemon}
             >
               Delete
             </button>
