+++ src/components/Detail/DetailPage.jsx
@@ -1,11 +1,13 @@
 import axios from "axios"
 import { useEffect, useState } from "react"
-import { useNavigate, useParams } from "react-router"
+import { useParams } from "react-router"
 import MissingNoImage from "../../assets/missingno.png"
 import Loading from "../../common/Loading"
 import PokemonEditForm from "./PokemonEditForm"
 import ConfirmDeletePopup from "./ConfirmDeletePopup"
 import { getJwt } from "../../utility/jwt"
+import BackHomeButton from "../../common/BackHomeButton"
+import PokemonDetails from "./PokemonDetails"
 
 const ENV = import.meta.env
 const API_URL =
@@ -13,7 +15,7 @@ const API_URL =
 
 const DetailPage = () => {
   const { id } = useParams() // parameter
-  const navigate = useNavigate()
+
   const [errorMessage, setErrorMessage] = useState("")
   const [isLoading, setIsLoading] = useState(true)
   const [pokemon, setPokemon] = useState({
@@ -25,7 +27,6 @@ const DetailPage = () => {
   const [boxShow, setBoxShow] = useState(false)
 
   const toggleBoxShow = () => setBoxShow(!boxShow)
-  console.log("boxshow = ", boxShow)
 
   const fetchDetails = async () => {
     setIsLoading(true)
@@ -45,62 +46,6 @@ const DetailPage = () => {
     }
   }
 
-  const updatePokemon = async (formData) => {
-    try {
-      if (!formData.name || !formData.type) {
-        throw new Error("Name and type are required")
-      }
-
-      if (pokemon) {
-        const jwt = getJwt()
-        const res = await axios.put(
-          `${API_URL}/pokemons/${id}`,
-          {
-            data: formData,
-          },
-          {
-            headers: {
-              Authorization: `Bearer ${jwt}`,
-            },
-          }
-        )
-        console.log("PUT res", res)
-      } else {
-        console.log("no data to update")
-      }
-    } catch (err) {
-      setErrorMessage(JSON.stringify(err.message))
-      console.log("Error")
-      console.error(err)
-    } finally {
-      await fetchDetails()
-    }
-  }
-
-  const deletePokemon = async () => {
-    try {
-      if (pokemon) {
-        const jwt = getJwt()
-        const res = await axios.delete(`${API_URL}/pokemons/${id}`, {
-          headers: {
-            Authorization: `Bearer ${jwt}`,
-          },
-        })
-        console.log("DELETE res", res)
-      } else {
-        console.log("no data to delete")
-      }
-      // go back to list page
-      navigate("/")
-    } catch (err) {
-      setErrorMessage(JSON.stringify(err.message))
-      console.log("Error")
-      console.error(err)
-    } finally {
-      await fetchDetails()
-    }
-  }
-
   useEffect(() => {
     const callFetchDetails = async () => {
       await fetchDetails()
@@ -113,40 +58,27 @@ const DetailPage = () => {
     <Loading />
   ) : (
     <div className="text-center">
-      <button
-        className="mt-5 underline text-slate-600 hover:text-slate-500"
-        onClick={() => navigate("/")}
-      >
-        Back to Home
-      </button>
-      <div className="block max-w-[320px] bg-slate-300 shadow-md rounded-xl py-4 px-5 mx-auto my-7">
-        <img
-          src={pokemon?.imageUrl}
-          alt={pokemon?.name}
-          className="object-contain w-full"
-        />
-        <p className="font-bold text-center">{pokemon?.name}</p>
-      </div>
-      <span className="px-4 py-3 text-white rounded-full bg-slate-400">
-        {pokemon?.type}
-      </span>
+      <BackHomeButton />
+      <PokemonDetails pokemon={pokemon} />
       {!!errorMessage && <p className="my-5 text-red-400">{errorMessage}</p>}
-      {boxShow && (
-        <ConfirmDeletePopup
-          showBox={boxShow}
-          onClose={toggleBoxShow}
-          onDelete={deletePokemon}
-        />
-      )}
+
+      <ConfirmDeletePopup
+        showBox={boxShow}
+        pokemon={pokemon}
+        setErrorMessage={setErrorMessage}
+        fetchDetails={fetchDetails}
+        onClose={toggleBoxShow}
+      />
+
       {pokemon && (
         <PokemonEditForm
           initialValues={pokemon}
-          handleSubmit={updatePokemon}
+          setErrorMessage={setErrorMessage}
+          fetchDetails={fetchDetails}
           handleDelete={toggleBoxShow}
         />
       )}
     </div>
   )
 }
-
 export default DetailPage
